* Style cleanup. * Changed quote from double to uint64.
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@42946 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/add-ons/mail_daemon/inbound_protocols/imap/imap_lib/IMAPFolders.h
@@ -39,7 +39,7 @@ public:
 			status_t			SubscribeFolder(const char* folder);
 			status_t			UnsubscribeFolder(const char* folder);
 
-			status_t			GetQuota(double& used, double& total);
+			status_t			GetQuota(uint64& used, uint64& total);
 private:
 			status_t			_GetAllFolders(StringList& folders);
 			status_t			_GetSubscribedFolders(StringList& folders);
